% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alma-in.R
\name{alma_in}
\alias{alma_in}
\title{Is \code{x} in the schedule?}
\usage{
alma_in(x, schedule)
}
\arguments{
\item{x}{\verb{[Date]}

A vector of dates.}

\item{schedule}{\verb{[schedule / rrule]}

A schedule or rrule.}
}
\description{
\code{alma_in()} checks if \code{x} is in the set of dates defined by the schedule.
}
