% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alma-seq.R
\name{alma_seq}
\alias{alma_seq}
\title{Generate events in a schedule}
\usage{
alma_seq(from, to, schedule, inclusive = TRUE)
}
\arguments{
\item{from, to}{\verb{[Date(1)]}

Dates defining the range to look for events in the \code{schedule}.}

\item{schedule}{\verb{[schedule / rrule]}

A schedule or rrule.}

\item{inclusive}{\verb{[logical(1)]}

If \code{from} or \code{to} are events, should they be included?}
}
\description{
\code{alma_seq()} generates all events in a schedule from \code{from} to \code{to}.
}
\examples{
on_12th <- monthly() \%>\% recur_on_mday(12)
on_monday <- weekly() \%>\% recur_on_wday("Monday")

# On the 12th of the month, or on Mondays
sch <- schedule() \%>\%
  sch_rrule(on_12th) \%>\%
  sch_rrule(on_monday)

alma_seq("2019-01-01", "2019-01-31", sch)

}
