% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-package.R
\docType{package}
\name{almanac-package}
\alias{almanac}
\alias{almanac-package}
\title{almanac: Tools for Working with Recurrence Rules}
\description{
Provides tools for defining recurrence rules and recurrence sets. Recurrence rules are a programmatic way to define a recurring event, like the first Monday of December. Multiple recurrence rules can be combined into larger recurrence sets. A full holiday and calendar interface is also provided that can generate holidays within a particular year, can detect if a date is a holiday, can respect holiday observance rules, and allows for custom holidays.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
