% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac4-package.R
\docType{package}
\name{almanac4-package}
\alias{almanac4}
\alias{almanac4-package}
\title{almanac4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac4}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac4/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
