% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdays.R
\docType{class}
\name{BDays-class}
\alias{BDays-class}
\title{Business Days}
\description{
BDays is an S4 class that contains a shift in business days, relative to
a schedule. The S4 nature of BDays is an implementation detail, and should
mostly be invisible to users.
}
\details{
BDays objects have two slots:
\itemize{
\item \code{.Data} \verb{[integer]}

The number of business days to shift by.
\item \code{schedule} \verb{[schedule]}

The schedule to shift relative to.
}

BDays place the number of business days in the \code{.Data} slot so that it can
"contain" (inherit from) \code{"integer"}. This makes it look like an integer
to vctrs and base methods like \code{length()}.
}
\examples{
new("BDays", 1L, schedule = schedule())
}
\keyword{internal}
