% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdays.R
\name{bdays}
\alias{bdays}
\alias{vec_arith.almanac_bdays}
\title{Create a business day period}
\usage{
bdays(n, schedule)

\method{vec_arith}{almanac_bdays}(op, x, y, ...)
}
\description{
Create a business day period
}
