% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcalendar.R
\name{cal_names}
\alias{cal_names}
\title{Calendar names}
\usage{
cal_names(x)
}
\arguments{
\item{x}{\verb{[rcalendar]}

An rcalendar.}
}
\value{
A character vector of holiday names.
}
\description{
\code{cal_names()} returns the names of the holidays in a calendar.
}
\examples{
x <- rcalendar(hol_christmas(), hol_new_years_day())
cal_names(x)
}
