% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendars}
\alias{calendars}
\alias{calendar_usa_federal}
\alias{calendar_usa_nyse}
\title{Calendars}
\usage{
calendar_usa_federal(since = "1970-01-01")

calendar_usa_nyse(since = "1970-01-01")
}
\arguments{
\item{since}{\verb{[Date(1)]}

The initial date to start the schedule from.}
}
\description{
The following generate \code{schedule} objects corresponding to calendars. A
calendar is composed of multiple holiday schedules generated from the
\verb{hldy_*()} functions.
}
\examples{
on_us_holidays <- calendar_usa_federal()

# Locate holidays
alma_search("2000-01-01", "2001-01-01", on_us_holidays)

# Adjust dates relative to holidays
on_weekends <- recur_on_weekends(weekly())
on_us_holidays_and_weekends <- add_rrule(on_us_holidays, on_weekends)

x <- as.Date("2019-08-30")
alma_step(x, 1, add_rrule(on_us_holidays, on_weekends))

}
