% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdays.R
\name{hidden-aliases}
\alias{hidden-aliases}
\alias{+,ANY,BDays-method}
\alias{+,BDays,ANY-method}
\alias{+,BDays,BDays-method}
\alias{+,BDays,Period-method}
\alias{+,Period,BDays-method}
\alias{-,ANY,BDays-method}
\alias{-,BDays,ANY-method}
\alias{-,BDays,BDays-method}
\alias{-,BDays,Period-method}
\alias{-,Period,BDays-method}
\alias{*,ANY,BDays-method}
\alias{*,BDays,ANY-method}
\alias{*,BDays,BDays-method}
\alias{*,BDays,Period-method}
\alias{*,Period,BDays-method}
\alias{/,ANY,BDays-method}
\alias{/,BDays,ANY-method}
\alias{/,BDays,BDays-method}
\alias{/,BDays,Period-method}
\alias{/,Period,BDays-method}
\alias{show,BDays-method}
\alias{xtfrm,BDays-method}
\title{Internal page for hidden aliases}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
