% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbundle.R
\name{is_rbundle}
\alias{is_rbundle}
\title{Is \code{x} a recurrence bundle?}
\usage{
is_rbundle(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"rbundle"}, otherwise \code{FALSE}.
}
\description{
\code{is_rbundle()} tests if \code{x} is a recurrence bundle.
}
\examples{
is_rbundle(rbundle())
}
