% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrule.R
\name{is_rrule}
\alias{is_rrule}
\title{Is \code{x} a recurrence rule?}
\usage{
is_rrule(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"rrule"}, otherwise \code{FALSE}.
}
\description{
\code{is_rrule()} tests if \code{x} is an rrule.
}
\examples{
is_rrule(daily())
}
