% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbundle-add.R
\name{rbundle-add}
\alias{rbundle-add}
\alias{add_rschedule}
\alias{add_rdate}
\alias{add_exdate}
\title{Add to an rbundle}
\usage{
add_rschedule(x, rschedule)

add_rdate(x, rdate)

add_exdate(x, exdate)
}
\arguments{
\item{x}{\verb{[rbundle]}

An rbundle to add to.}

\item{rschedule}{\verb{[rschedule]}

An rschedule, such as an rrule or rbundle.}

\item{rdate}{\verb{[Date]}

Dates to forcibly include in the rbundle.}

\item{exdate}{\verb{[Date]}

Dates to forcibly exclude from the rbundle.}
}
\description{
\itemize{
\item \code{add_rschedule()} adds an rschedule to an rbundle. This can be another
rrule or another rbundle.
\item \code{add_rdate()} adds a rdate to an rbundle. rdates are singular
special cased dates that are forcibly included in the recurrence set.
\item \code{add_exdate()} adds an exdate to a rbundle. exdates are singular
special cased dates that are forcibly excluded from the recurrence set.
}
}
\details{
In terms of priority:
\itemize{
\item An exdate will never be included.
\item A rdate will always be included if it is not also an exdate.
\item An event generated from an rschedule will always be included if it is not
also an exdate.
}

Combining two rrules into the same rbundle is a way of joining them using
an "or" condition. For example, joining the following recurrence rules
would translate to "the 5th day of the month or any Tuesday".\preformatted{on_5th_of_the_month <- monthly() \%>\%
  on_mday(5)

on_tuesday <- weekly() \%>\%
  on_wday("Tuesday")

rbundle() \%>\%
  add_rschedule(on_5th_of_the_month) \%>\%
  add_rschedule(on_tuesday)
}
}
\examples{
on_thanksgiving <- yearly() \%>\%
  recur_on_wday("Thurs", 4) \%>\%
  recur_on_ymonth("Nov")

on_christmas <- yearly() \%>\%
  recur_on_mday(25) \%>\%
  recur_on_ymonth("Dec")

on_labor_day <- monthly() \%>\%
  recur_on_ymonth("Sep") \%>\%
  recur_on_wday("Mon", 1)

rb <- rbundle() \%>\%
  add_rschedule(on_thanksgiving) \%>\%
  add_rschedule(on_christmas) \%>\%
  add_rschedule(on_labor_day)

# Thanksgiving, Christmas, or Labor Day
alma_search("2019-01-01", "2021-01-01", rb)

# Except Labor Day in 2019
rb2 <- add_exdate(rb, "2019-09-02")

alma_search("2019-01-01", "2021-01-01", rb2)

}
