% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbundle.R
\name{rbundle}
\alias{rbundle}
\title{Create a new recurrence bundle}
\usage{
rbundle()
}
\value{
An empty rbundle.
}
\description{
Often, a single recurrence rule created from a base rule like \code{monthly()}
will be sufficient. However, more complex rules can be constructed
by combining simple rules into a \emph{recurrence bundle}.

\code{rbundle()} creates a new empty recurrence bundle.
\itemize{
\item Add recurrence rules or other recurrence bundles with \code{\link[=add_rschedule]{add_rschedule()}}.
\item Forcibly include dates with \code{\link[=add_rdate]{add_rdate()}}.
\item Forcibly exclude dates with \code{\link[=add_exdate]{add_exdate()}}.
}
}
\examples{
rbundle()

on_weekends <- weekly() \%>\%
  recur_on_weekends()

rbundle() \%>\%
  add_rschedule(on_weekends)
}
\seealso{
\code{\link[=add_rschedule]{add_rschedule()}}
}
