% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur-until.R
\name{recur_until}
\alias{recur_until}
\title{Control the final date of the recurrence}
\usage{
recur_until(x, until)
}
\arguments{
\item{x}{\verb{[rrule]}

A recurrence rule.}

\item{until}{\verb{[Date(1)]}

The final date boundary of the recurrence.}
}
\description{
\code{recur_until()} controls the final date of the recurrence set. If not set,
the recurrence will continue indefinitely. \code{recur_until()}
is mutually exclusive with \code{\link[=recur_for_count]{recur_for_count()}}.
}
\details{
Remember that the \code{until} date must be after the \code{since} date! Adjust it as
necessary to get your desired results.

The \code{until} date is only included as an event if it passes all of the checks
set by the recurrence rule. Otherwise, it is just used as a boundary.
}
\examples{
# Using the default `since` date
daily_since_epoch_limited <- daily() \%>\% recur_until("1970-01-05")

alma_seq("1969-12-31", "1970-01-25", daily_since_epoch_limited)

# Changing the `since` date
daily_since_2019_limited <- daily(since = "2019-01-01") \%>\%
  recur_until("2019-01-05")

alma_seq("2018-12-31", "2019-01-25", daily_since_2019_limited)

}
