% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{days}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate]{days}}}
}}

