% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr-with-week-start.R
\name{rr_with_week_start}
\alias{rr_with_week_start}
\title{Control the start of the week}
\usage{
rr_with_week_start(x, wday)
}
\arguments{
\item{x}{\code{[rrule]}

A recurrence rule.}

\item{wday}{\code{[integer(1) / character(1)]}

Day of the week to start the week on. Must be an integer value in
\code{[1, 7]}, with \code{1 = Monday} and \code{7 = Sunday}. This is also allowed to be
a full weekday string like \code{"Tuesday"}, or an abbreviation like \code{"Tues"}.}
}
\description{
\code{rr_with_week_start()} controls the week day that represents the start of the
week. This is important for rules that use \code{\link[=rr_on_yweek]{rr_on_yweek()}}. See that page
for examples.

\emph{The default day of the week to start on is Monday.}
}
