% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch-in.R
\name{sch_in}
\alias{sch_in}
\title{Is \code{x} in the schedule?}
\usage{
sch_in(x, schedule)
}
\arguments{
\item{x}{\code{[Date]}

A vector of dates.}

\item{schedule}{\code{[schedule / rrule]}

A schedule or rrule.}
}
\description{
\code{sch_in()} checks if \code{x} is in the set of dates defined by the schedule.
}
