% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch-range.R
\name{sch_next}
\alias{sch_next}
\alias{sch_previous}
\title{Generate the next or previous event}
\usage{
sch_next(x, schedule, inclusive = FALSE)

sch_previous(x, schedule, inclusive = FALSE)
}
\arguments{
\item{x}{\code{[Date(1)]}

The date to start the search from.}

\item{schedule}{\code{[schedule / rrule]}

A schedule or rrule.}

\item{inclusive}{\code{[logical(1)]}

If \code{from} or \code{to} are events, should they be included?}
}
\description{
\itemize{
\item \code{sch_next()} generates the next event after \code{x}.
\item \code{sch_previous()} generates the previous event before \code{x}.
}
}
\examples{
library(magrittr)

on_12th <- monthly() \%>\% rr_on_mday(12)
on_monday <- weekly() \%>\% rr_on_wday("Monday")

# On the 12th of the month, or on Mondays
sch <- schedule() \%>\%
  sch_add_rrule(on_12th) \%>\%
  sch_add_rrule(on_monday)

sch_next("2019-01-01", sch)

sch_previous("2019-01-01", sch)

}
