% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch-range.R
\name{sch_seq}
\alias{sch_seq}
\title{Generate events in a schedule}
\usage{
sch_seq(from, to, schedule, inclusive = TRUE)
}
\arguments{
\item{from, to}{\code{[Date(1)]}

Dates defining the range to look for events in the \code{schedule}.}

\item{schedule}{\code{[schedule / rrule]}

A schedule or rrule.}

\item{inclusive}{\code{[logical(1)]}

If \code{from} or \code{to} are events, should they be included?}
}
\description{
\code{sch_seq()} generates all events in a schedule from \code{from} to \code{to}.
}
\examples{
library(magrittr)

on_12th <- monthly() \%>\% rr_on_mday(12)
on_monday <- weekly() \%>\% rr_on_wday("Monday")

# On the 12th of the month, or on Mondays
sch <- schedule() \%>\%
  sch_add_rrule(on_12th) \%>\%
  sch_add_rrule(on_monday)

sch_seq("2019-01-01", "2019-01-31", sch)

}
