% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{Create a new schedule}
\usage{
schedule()
}
\description{
Often, a single recurrence rule created from a base rule like \code{monthly()}
will be sufficient. However, more complex rules can be constructed
by combining simple rules into a \emph{schedule}.

\code{schedule()} creates a new empty schedule. Add rules to the schedule with
\code{\link[=sch_rrule]{sch_rrule()}}.
}
\examples{
schedule()

sch_rrule(schedule(), monthly())

}
\seealso{
\code{\link[=sch_rrule]{sch_rrule()}}
}
