% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdays.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_arith.BDays}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{BDays}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string.}

\item{x, y}{Objects.}

\item{...}{Used to pass along error message information.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow rray objects to
work with vctrs.
}
