% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alma-events.R
\name{alma_events}
\alias{alma_events}
\title{Get all events}
\usage{
alma_events(schedule)
}
\arguments{
\item{schedule}{\verb{[schedule / rrule]}

A schedule or rrule.}
}
\value{
A \code{Date} vector of events.
}
\description{
\code{alma_events()} retrieves all of the events defined by the schedule's
recurrence rules.
}
\examples{
rrule <- daily(since = "1970-01-01", until = "1970-01-05")

alma_events(rrule)

rrule_weekly <- weekly(since = "1970-01-01") \%>\%
  recur_for_count(5)

sch <- schedule() \%>\%
  sch_rrule(rrule) \%>\%
  sch_rrule(rrule_weekly)

alma_events(sch)
}
