% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alma-in.R
\name{alma_in}
\alias{alma_in}
\title{Is \code{x} in the recurrence set?}
\usage{
alma_in(x, rbundle)
}
\arguments{
\item{x}{\verb{[Date]}

A vector of dates.}

\item{rbundle}{\verb{[rbundle / rrule]}

An rbundle or rrule.}
}
\description{
\code{alma_in()} checks if \code{x} is in the recurrence set of dates defined by the
rbundle.
}
\examples{
rrule <- weekly() \%>\%
  recur_on_wday("Thursday")

# A Thursday and Friday
x <- as.Date("1970-01-01") + 0:1

alma_in(x, rrule)

# Every month, on the 2nd day of the month
rrule2 <- monthly() \%>\%
  recur_on_mday(2)

# Make a larger rbundle made of multiple rules
rb <- rbundle() \%>\%
 add_cacher(rrule) \%>\%
 add_cacher(rrule2)

alma_in(x, rb)
}
