% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-package.R
\docType{package}
\name{almanac-package}
\alias{almanac}
\alias{almanac-package}
\title{almanac: Tools for Working with Recurrence Rules}
\description{
Provides tools for defining recurrence rules and recurrence bundles. Recurrence rules are a programmatic way to define a recurring event, like the first Monday of December. Multiple recurrence rules can be combined into larger recurrence bundles. Together, these provide a system for adjusting and generating sequences of dates while simultaneously skipping over dates in a recurrence bundle's event set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
