% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-package.R
\docType{package}
\name{almanac-package}
\alias{almanac}
\alias{almanac-package}
\title{almanac: Tools for Adjusting and Generating Dates Using a Grammar of Schedules}
\description{
Provides tools for defining events and schedules. These provide a 
    system for generating dates in a schedule, adjusting them relative to a 
    schedule, and determining if they are included in a schedule.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
