% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{almanac-vctrs-compat}
\alias{almanac-vctrs-compat}
\alias{vec_arith.almanac_stepper}
\alias{vec_ptype2.almanac_stepper.almanac_stepper}
\alias{vec_cast.almanac_stepper.almanac_stepper}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{almanac_stepper}(op, x, y, ...)

\method{vec_ptype2}{almanac_stepper.almanac_stepper}(x, y, ..., x_arg = "", y_arg = "")

\method{vec_cast}{almanac_stepper.almanac_stepper}(x, to, ..., x_arg = "", to_arg = "")
}
\arguments{
\item{op}{An arithmetic operator as a string.}

\item{x, y, to}{Objects.}

\item{...}{Used to pass along error message information.}

\item{x_arg, y_arg, to_arg}{Used to pass along error message information.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow stepper objects to
work with vctrs.
}
\examples{
library(vctrs)
vec_arith("+", as.Date("2019-01-04"), workdays(1))
}
\keyword{internal}
