% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-deprecated.R
\name{deprecated-recur}
\alias{deprecated-recur}
\alias{recur_on_mday}
\title{Deprecated recurrence helpers}
\usage{
recur_on_mday(x, mday)
}
\arguments{
\item{x}{\verb{[rrule]}

A recurrence rule.}

\item{mday}{\verb{[integer]}

The days of the month on which to recur. Negative values are allowed,
which specify \code{n} days from the end of the month.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{recur_on_mday()} is deprecated as of almanac 1.0.0 in favor of
\code{\link[=recur_on_day_of_month]{recur_on_day_of_month()}}.
}
}
\keyword{internal}
