% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{holidays}
\alias{holidays}
\alias{hldy_washington_birthday}
\alias{hldy_memorial_day}
\alias{hldy_labor_day}
\alias{hldy_columbus_day}
\alias{hldy_veterans_day}
\alias{hldy_new_years_day}
\alias{hldy_martin_luther_king_jr_day}
\alias{hldy_independence_day}
\alias{hldy_thanksgiving}
\alias{hldy_christmas}
\alias{hldy_easter}
\alias{hldy_good_friday}
\alias{hldy_easter_monday}
\alias{hldy_july_fourth}
\alias{hldy_lincoln_birthday}
\alias{hldy_election_day}
\title{Holidays}
\usage{
hldy_washington_birthday(since = "1970-01-01")

hldy_memorial_day(since = "1970-01-01")

hldy_labor_day(since = "1970-01-01")

hldy_columbus_day(since = "1970-01-01", until = NULL)

hldy_veterans_day(since = "1970-01-01")

hldy_new_years_day(since = "1970-01-01")

hldy_martin_luther_king_jr_day(since = "1970-01-01")

hldy_independence_day(since = "1970-01-01")

hldy_thanksgiving(since = "1970-01-01")

hldy_christmas(since = "1970-01-01")

hldy_easter(since = "1970-01-01", until = NULL)

hldy_good_friday(since = "1970-01-01", until = NULL)

hldy_easter_monday(since = "1970-01-01", until = NULL)

hldy_july_fourth(since = "1970-01-01")

hldy_lincoln_birthday(since = "1970-01-01", until = NULL)

hldy_election_day(since = "1970-01-01", until = NULL)
}
\arguments{
\item{since}{\code{[Date(1)]}

The initial date to start the schedule from.}
}
\description{
The following generate \code{schedule} objects corresponding to commonly used
holidays.
}
\examples{
on_thanksgiving <- hldy_thanksgiving()

sch_seq("2000-01-01", "2019-01-01", on_thanksgiving)

}
