% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbundle.R
\name{new_rbundle}
\alias{new_rbundle}
\title{Constructor for an rbundle}
\usage{
new_rbundle(rschedules = list(), rdates = new_date(), exdates = new_date())
}
\arguments{
\item{rschedules}{\verb{[list]}

A list of rschedules.}

\item{rdates}{\verb{[Date]}

A vector of dates to forcibly include in the recurrence set.}

\item{exdates}{\verb{[Date]}

A vector of dates to forcibly exclude from the recurrence set.}
}
\value{
A new rbundle.
}
\description{
\code{new_rbundle()} is a developer focused tool that is not required for normal
usage of almanac. It constructs a new rbundle directly from a list of
existing rschedules.
}
\examples{
new_rbundle()

x <- daily()
y <- weekly()

rschedules <- list(x, y)

new_rbundle(rschedules)
}
