% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rschedule.R
\name{new_rschedule}
\alias{new_rschedule}
\alias{rschedule_events}
\title{Create a new rschedule}
\usage{
new_rschedule(data, class)

rschedule_events(x)
}
\arguments{
\item{data}{\verb{[named list]}

A named list of data fields.}

\item{class}{\verb{[character]}

A required subclass.}

\item{x}{\verb{[rschedule subclass]}

An object that subclasses rschedule.}
}
\value{
For \code{new_rschedule()}, a new rschedule subclass.

For \code{rschedule_events()}, a Date vector of events.
}
\description{
\code{new_rschedule()} is a developer focused tool that is not required for
normal usage of almanac. It is only exported to allow other packages
to construct new rschedule objects that work with almanac functions
prefixed with \verb{alma_*()}, like \code{\link[=alma_in]{alma_in()}}.

\code{rschedule_events()} is a generic function that rschedule subclasses must
provide a method for. \code{rschedule_events()} should return a Date vector
containing the complete ordered set of events in the recurrence set of
that rschedule.
}
\details{
An rschedule is an abstract class that rrule and rbundle both inherit from.
The sole functionality of rschedule classes is to provide a method for
\code{rschedule_events()}.
}
\examples{
events <- as.Date("1970-01-01")

static <- new_rschedule(
  list(events = events),
  class = "static_rschedule"
)

rschedule_events.static_rschedule <- function(x) {
  x$events
}

alma_events(static)

alma_in(events, static)
alma_in(events + 1, static)
}
