% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbundle.R
\name{rbundle}
\alias{rbundle}
\title{Create a new recurrence bundle}
\usage{
rbundle()
}
\value{
An empty rbundle.
}
\description{
Often, a single recurrence rule created from a base rule like \code{monthly()}
will be sufficient. However, more complex rules can be constructed
by combining simple rules into a \emph{recurrence bundle}.

\code{rbundle()} creates a new empty recurrence bundle. Add recurrence rules to
the bundle with \code{\link[=add_cacher]{add_cacher()}}. Add required dates with \code{\link[=add_rdate]{add_rdate()}}.
}
\examples{
rbundle()
add_cacher(rbundle(), monthly())
}
\seealso{
\code{\link[=add_cacher]{add_cacher()}}
}
