% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur-on-ymonth.R
\name{recur_on_ymonth}
\alias{recur_on_ymonth}
\title{Recur on a month of the year}
\usage{
recur_on_ymonth(x, ymonth)
}
\arguments{
\item{x}{\code{[rrule]}

A recurrence rule.}

\item{ymonth}{\code{[integer / character]}

Months of the year the mark as events. Integer values must be between
\code{[1, 12]}. This can also be a full month string like \code{"November"}, or an
abbreviation like \code{"Nov"}.}
}
\description{
\code{recur_on_ymonth()} recurs on a specific month of the year.
}
\examples{
library(magrittr)

# There is a big difference between adding this rule to a `yearly()`
# or `monthly()` frequency, and a `daily()` frequency.

# Limit from every day to every day in February
on_feb_daily <- daily() \%>\% recur_on_ymonth("Feb")

# Limit from 1 day per month to 1 day in February
on_feb_monthly <- monthly() \%>\% recur_on_ymonth("Feb")

start <- "1999-01-01"
end <- "2001-01-01"

sch_seq(start, end, on_feb_daily)

sch_seq(start, end, on_feb_monthly)

}
